# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Digital Product Passport (DPP) Service API
  description: >-
    REST/HTTPS API for the Digital Product Passport, derived from the
    CEN/CENELEC JTC 24 draft standards:
      * prEN 18222 — APIs for the product passport lifecycle management
        (endpoints, HTTP verbs, parameters, status codes — Tables 1–19)
      * prEN 18223 — System interoperability (payload / semantic data model)
      * prEN 18216 — Data exchange protocols (TLS, JSON/XML/JSON-LD/HTML, error handling)
      * prEN 18239 — Access rights, information security, business confidentiality
      * prEN 18219 — Unique identifiers (incl. W3C DIDs — relevant for DID FlexCo)

    NOTE: The source standards are 2025 drafts (Entwurf / CEN Enquiry) and are
    technology-neutral. This specification is one concrete REST realisation of
    the abstract methods and MUST be re-validated against the final published
    standards. semanticId base URI `https://jtc24/...` is a placeholder from the draft.

    IDENTIFIERS IN PATHS: `dpp_id` and `product_id` are URIs/URLs (prEN 18219)
    and contain `:`, `/` and `.`. They MUST be percent-encoded into a single
    path segment, e.g.
    `https://dpp.example.org/01/09520123456788/8546` becomes
    `https%3A%2F%2Fdpp.example.org%2F01%2F09520123456788%2F8546`.
    The service disables content-type suffix routing (`.json`), so dots inside
    an identifier are preserved.
  version: 0.1.0
  contact:
    name: DPP Service (DID FlexCo)
  license:
...
servers:
...
security:
  - bearerAuth: []   # Public read data is accessible unauthenticated (prEN 18239, §6.2); controlled data requires a token.
paths:
...
  /registerDPP:
    post:
      tags: [Registry API]
      summary: PostNewDPPToRegistry
      description: >-
        Registers a new DPP at the EC registry by providing product, backup and
        operator identifiers. Served by the EC registry server; returns a unique
        registry identifier (prEN 18222 §5.2, Table 8).
        NOTE: The concrete EC registry endpoint/schema is defined by EU
        implementing acts — this operation models the client-facing contract.
      operationId: postNewDPPToRegistry
      x-semanticId: https://jtc24/dpp/API/PostNewDPPToRegistry/1/0
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegistryRegistration' }
      responses:
        '201':
          description: SuccessCreated — DPP registered.
          content:
            application/json:
              schema:
                type: object
                required: [registryIdentifier]
                properties:
                  registryIdentifier: { $ref: '#/components/schemas/Identifier' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/NotAuthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '502': { $ref: '#/components/responses/BadGateway' }

  ###########################################################################
  # 6  Fine Granular API Operations of the Life Cycle API
  ###########################################################################
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        OAuth 2.0 / OpenID Connect access token (JWT). Public data is readable
        without authentication; controlled data requires an authenticated actor
        with an appropriate role (authority, refurbisher, consumer, ...) per prEN 18239.
...
  responses:
    BadRequest:
      description: ClientErrorBadRequest — bad or malformed request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Result' }
    NotAuthorized:
      description: ClientNotAuthorized — wrong or missing authorization credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Result' }
...
    Conflict:
      description: ClientResourceConflict — resource already exists.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Result' }
...
    BadGateway:
      description: ServerErrorBadGateway — bad gateway.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Result' }

  ###########################################################################
  # Schemas
  ###########################################################################
  schemas:
    Identifier:
      type: string
      description: Globally unique identifier; SHOULD be a URI/URL (see prEN 18219, incl. W3C DID schemes).
      example: "https://dpp.example.org/01/09520123456788/21/12345"

    # ---- prEN 18222 §7: Status / Result / Message ------------------------
    StatusCode:
      type: string
      description: Generic status code of the response (prEN 18222 Table 16).
      enum:
        - Success
        - SuccessCreated
        - SuccessAccepted
        - SuccessNoContent
        - ClientErrorBadRequest
        - ClientNotAuthorized
        - ClientForbidden
        - ClientMethodNotAllowed
        - ClientErrorResourceNotFound
        - ClientResourceConflict
        - ServerInternalError
        - ServerErrorBadGateway
    MessageTypeEnum:
      type: string
      enum: [Info, Warning, Error, Exception]
      description: The message type (prEN 18222 Table 15).
    Message:
      type: object
      description: A message with more information for the requester (prEN 18222 Table 14).
      required: [messageType, text]
      properties:
        messageType: { $ref: '#/components/schemas/MessageTypeEnum' }
        text: { type: string }
        code:
          type: string
          description: Technology-dependent status or error code.
        correlationId:
          type: string
          description: Identifier to relate several result messages across systems.
        timestamp:
          type: string
          format: date-time
    Result:
      type: object
      description: Result object returned on failed method execution (prEN 18222 Table 13).
      properties:
        statusCode: { $ref: '#/components/schemas/StatusCode' }
        message:
          type: array
          items: { $ref: '#/components/schemas/Message' }

    # ---- prEN 18223 §4.1: DPP semantic model -----------------------------
...
    RegistryRegistration:
      type: object
      description: Information about the operator, backup operator and product to be registered.
      required: [ProductID, OperatorID]
      properties:
        ProductID: { $ref: '#/components/schemas/Identifier' }
        BackupID:
          allOf: [{ $ref: '#/components/schemas/Identifier' }]
          description: Identifier of the backup operator.
        OperatorID: { $ref: '#/components/schemas/Identifier' }